backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
pyjwt==2.12.1
chromadb==0.4.17
pypdf==3.17.1
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytesseract==0.3.10
Pillow==10.1.0

